Rename file to CamelCase(PluginBase)
