+++ packages/core/src/PluginSupport/PluginBase.ts
