+++ packages/core/src/PluginSupport/PostEffectBase.ts
@@ -1,5 +1,5 @@
 import { EffectPreRenderContext, EffectRenderContext } from '..'
-import PluginBase from './plugin-base'
+import PluginBase from './PluginBase'
 import { TypeDescriptor } from './TypeDescriptor'
 
 export interface EffectPluginClass {
