+++ packages/core/src/PluginSupport/index.ts
@@ -1,4 +1,4 @@
-export { default as PluginBase } from './plugin-base'
+export { default as PluginBase } from './PluginBase'
 export { default as PostEffectBase } from './PostEffectBase'
 export { default as PluginRegistry } from './plugin-registry'
 export { default as TypeDescriptor } from './TypeDescriptor'
